ajout de ce qu'il y a dans le fichier clienttest vers le main

+++ lib/controller/wavecontroller.dart
@@ -14,8 +14,8 @@ class WaveController {
       */
 
       default: {
-        for(int i = 0; i < wave*10; i++){
-          listEnemy.add(EnemyWidget.enemyWidgetRandom(minX, maxX, minY, maxY, 2, images));
+        for(int i = 0; i < wave.toDouble()*10; i++){
+          listEnemy.add(EnemyWidget.enemyWidgetRandom(minX, maxX, minY, maxY, 2.toInt(), images));
         }
       }
       break;
